Рефактор UserDao и IUserDao Рефактор DB закончен

+++ src/main/java/ru/innopolis/stc9/susev/db/dao/UserDao.java
@@ -14,55 +14,45 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class UserDao implements IUserDao {
-
-    public static final String COLUMN_ID = "id";
-    public static final String COLUMN_LOGIN = "login";
-    public static final String COLUMN_PASSWORD = "password";
-    public static final String COLUMN_ROLE = "role";
-    public static final String COLUMN_FULL_NAME = "fullName";
-
     private ConnectionManager conManager = ConnectionManagerJDBCImpl.getInstance();
 
     @Nullable
     @Override
     public User getUserByLogin(String login) throws SQLException {
-        User result = null;
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM users WHERE login=?");
-        preparedStatement.setString(1, login);
-        ResultSet resultSet = preparedStatement.executeQuery();
-        if (resultSet.next()) {
-            IRoleDao roleDao = new RoleDao();
-            Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
-            result = new User(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_LOGIN),
-                    resultSet.getString(COLUMN_PASSWORD),
-                    role,
-                    resultSet.getString(COLUMN_FULL_NAME));
+        if (login == null || login.isEmpty()) {
+            return null;
         }
-        connection.close();
-        return result;
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM users WHERE login=?");
+            preparedStatement.setString(1, login);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            if (resultSet.next()) {
+                IRoleDao roleDao = new RoleDao();
+                Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
+                return getUserFromDb(resultSet, role);
+            }
+            return null;
+        }
     }
+
 
     @Nullable
     @Override
-    public User getUserById(String id) throws SQLException {
-        User result = null;
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM users WHERE id=?");
-        preparedStatement.setInt(1, Integer.parseInt(id));
-        ResultSet resultSet = preparedStatement.executeQuery();
-        if (resultSet.next()) {
-            IRoleDao roleDao = new RoleDao();
-            Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
-            result = new User(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_LOGIN),
-                    resultSet.getString(COLUMN_PASSWORD),
-                    role,
-                    resultSet.getString(COLUMN_FULL_NAME));
+    public User getUserById(int id) throws SQLException {
+        if (id == 0) {
+            return null;
         }
-        connection.close();
-        return result;
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM users WHERE id=?");
+            preparedStatement.setInt(1, id);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            if (resultSet.next()) {
+                IRoleDao roleDao = new RoleDao();
+                Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
+                return getUserFromDb(resultSet, role);
+            }
+            return null;
+        }
     }
 
     @Override
@@ -70,31 +60,25 @@ public class UserDao implements IUserDao {
         if (user == null) {
             return false;
         }
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("insert into users(login, password, role, fullName) " +
-                "values (?, ?, ?, ?)");
-        statement.setString(1, user.getLogin());
-        statement.setString(2, user.getPasswordHash());
-        statement.setInt(3, user.getRole().getId());
-        statement.setString(4, user.getFullName());
-
-        boolean execute = statement.execute();
-        connection.close();
-        return execute;
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement("insert into users(login, password, role, fullName) " +
+                    "values (?, ?, ?, ?)");
+            setParamsIntoStatement(statement, user);
+            return statement.execute();
+        }
     }
+
 
     @Override
     public boolean deleteUser(User user) throws SQLException {
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("delete from users where id = ?");
-        statement.setInt(1, user.getId());
-        boolean execute = statement.execute();
-        connection.close();
-        return execute;
+        if (user == null) {
+            return false;
+        }
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement("delete from users where id = ?");
+            statement.setInt(1, user.getId());
+            return statement.execute();
+        }
     }
 
     @Override
@@ -102,90 +86,79 @@ public class UserDao implements IUserDao {
         if (user == null) {
             return false;
         }
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("update users " +
-                "set login = ?, password = ?, role = ?, fullName = ?" +
-                "where id = ?");
-        statement.setString(1, user.getLogin());
-        statement.setString(2, user.getPasswordHash());
-        statement.setInt(3, user.getRole().getId());
-        statement.setString(4, user.getFullName());
-        statement.setInt(5, user.getId());
-
-        statement.executeUpdate();
-        connection.close();
-        return true;
-    }
-
-    @Override
-    public Role getRoleByLogin(String login) throws SQLException {
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("select r.id, r.role from users " +
-                "inner join role r on r.id = users.role " +
-                "where users.login = ?");
-        statement.setString(1, login);
-        ResultSet set = statement.executeQuery();
-        if (set.next()) {
-            return new Role(
-                    set.getInt(COLUMN_ID),
-                    set.getString(COLUMN_ROLE)
-            );
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement("update users " +
+                    "set login = ?, password = ?, role = ?, fullName = ?" +
+                    "where id = ?");
+            setParamsIntoStatement(statement, user);
+            statement.setInt(5, user.getId());
+            statement.executeUpdate();
+            return true;
         }
-        connection.close();
-        return null;
     }
+
 
     @Override
     public boolean deleteUserById(int idUser) throws SQLException {
-        Connection connection;
-        connection = conManager.getConnection();
-        PreparedStatement statement = null;
-        statement = connection.prepareStatement("delete from users where id = ?");
-        statement.setInt(1, idUser);
-        boolean execute = statement.execute();
-        connection.close();
-        return execute;
+        if (idUser == 0) {
+            return false;
+        }
+        try (Connection connection = conManager.getConnection()) {
+            PreparedStatement statement = connection.prepareStatement("delete from users where id = ?");
+            statement.setInt(1, idUser);
+            return statement.execute();
+        }
     }
 
     @Override
     public List<User> getUsers(int roleId) throws SQLException {
-        List<User> result = new ArrayList<>();
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        String sql = "select * from users where users.role = " + roleId + ";";
-        PreparedStatement preparedStatement = connection.prepareStatement(sql);
-        ResultSet resultSet = preparedStatement.executeQuery();
-        while (resultSet.next()) {
-            IRoleDao roleDao = new RoleDao();
-            Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
-            result.add(new User(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_LOGIN),
-                    resultSet.getString(COLUMN_PASSWORD),
-                    role,
-                    resultSet.getString(COLUMN_FULL_NAME)));
+        if (roleId == 0) {
+            return null;
         }
-        connection.close();
-        return result;
+        try (Connection connection = conManager.getConnection()) {
+            List<User> result = new ArrayList<>();
+            String sql = "select * from users where users.role = " + roleId + ";";
+            PreparedStatement preparedStatement = connection.prepareStatement(sql);
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                IRoleDao roleDao = new RoleDao();
+                Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
+                result.add(getUserFromDb(resultSet, role));
+            }
+            return result;
+        }
     }
 
+    @Override
     public List<User> getUsers() throws SQLException {
-        List<User> result = new ArrayList<>();
-        Connection connection = ConnectionManagerJDBCImpl.getInstance().getConnection();
-        PreparedStatement preparedStatement = connection.prepareStatement("select * from users;");
-        ResultSet resultSet = preparedStatement.executeQuery();
-        while (resultSet.next()) {
-            IRoleDao roleDao = new RoleDao();
-            Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
-            result.add(new User(resultSet.getInt(COLUMN_ID),
-                    resultSet.getString(COLUMN_LOGIN),
-                    resultSet.getString(COLUMN_PASSWORD),
-                    role,
-                    resultSet.getString(COLUMN_FULL_NAME)));
+        try (Connection connection = conManager.getConnection()) {
+            List<User> result = new ArrayList<>();
+            PreparedStatement preparedStatement = connection.prepareStatement("select * from users;");
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                IRoleDao roleDao = new RoleDao();
+                Role role = roleDao.getRoleById(resultSet.getInt(COLUMN_ROLE));
+                result.add(getUserFromDb(resultSet, role));
+            }
+            return result;
         }
-        connection.close();
-        return result;
+    }
+
+    /**
+     * Get user from DB
+     */
+    private User getUserFromDb(ResultSet resultSet, Role role) throws SQLException {
+        return new User(resultSet.getInt(COLUMN_ID), resultSet.getString(COLUMN_LOGIN),
+                resultSet.getString(COLUMN_PASSWORD), role, resultSet.getString(COLUMN_FULL_NAME));
+    }
+
+    /**
+     * Set params from user into statement
+     */
+    private void setParamsIntoStatement(PreparedStatement statement, User user) throws SQLException {
+        statement.setString(1, user.getLogin());
+        statement.setString(2, user.getPasswordHash());
+        statement.setInt(3, user.getRole().getId());
+        statement.setString(4, user.getFullName());
     }
 }
